+++ packages/frontend/src/features/editor/hooks/use-editor-builder.tsx
@@ -11,13 +11,21 @@ import TableMenu from "../extensions/table/table-menu";
 import { EventHelper } from "../components/event-helper";
 import { FormattingHelper } from "../components/formatting-helper";
 import { DragHandleExtension } from "../extensions/drag-handle";
+import { Placeholder } from "@tiptap/extensions";
+import { useTranslation } from "react-i18next";
 
 export default function useEditorBuilder() {
   const { imageUploadConfig, codeBlockIndentSize, commandItems } = use(
     DarkwriteEditorContext,
   );
-
+  const { t } = useTranslation();
   const isDragging = useRef(false);
+
+  const placeholder = Placeholder.configure({
+    includeChildren: true,
+    placeholder: () => t("editor.placeholder"),
+    showOnlyCurrent: true,
+  });
 
   const extensions = useMemo(
     () => [
@@ -29,8 +37,9 @@ export default function useEditorBuilder() {
           items: () => commandItems,
         },
       }),
+      placeholder,
     ],
-    [codeBlockIndentSize, imageUploadConfig, commandItems],
+    [imageUploadConfig, placeholder, codeBlockIndentSize, commandItems],
   );
 
   const children = useMemo(